fix: renamed directories, renamed script, rectified message for successful request, renamed and seperated util functions
