+++ utils/database.js
@@ -1,0 +1,60 @@
+require('dotenv').config();
+const { supabaseClient } = require('../config/supabaseConfig');
+
+
+const sendMessage = (res, statusCode, errorMsg, msgContent, data) => {
+  const successCode = [200, 201, 202, 204, 206];
+  if (successCode.includes(statusCode)) {
+    res
+      .status(statusCode)
+      .json({ status: true, message: msgContent, data: data || {} });
+
+    return;
+  }
+  if (!statusCode || statusCode === 500) {
+    msgContent = 'Internal Server Error.';
+  }
+
+  res
+    .status(statusCode)
+    .json({ error: errorMsg, message: msgContent, data: data || {} });
+};
+
+const queryDatabase = async (tablename, column, value) => {
+  try {
+    const { data: existingRecord, error: Error } = await supabaseClient
+      .from(tablename)
+      .select()
+      .eq(column, value);
+
+    if (Error) {
+      console.error('Error fetching from database...', Error.message);
+      throw Error;
+    }
+
+    return existingRecord;
+  } catch (error) {
+    console.error('Error from queryDatabase: ', error.message);
+    throw error;
+  }
+};
+
+const insertIntoDatabase = async (tablename, data) => {
+  try {
+    const { error: Error } = await supabaseClient.from(tablename).insert(data);
+
+    if (Error) {
+      console.error('Error inserting into database', Error.message);
+      throw Error;
+    }
+  } catch (error) {
+    console.log('Error from insertIntoDatabase', error.message);
+    throw error;
+  }
+};
+
+module.exports = {
+  sendMessage,
+  queryDatabase,
+  insertIntoDatabase,
+};
